[gobdk] MvP-4858 widen the C length ABI to uint64_t and add the ABI error domain
Every buffer length crossing the gobdk C ABI was narrowed to a C int with no
range check, so an extended transaction between 2^31 and 2^32-1 bytes arrived
as a negative length. For the extended-transaction argument that produced an
empty span, a parse failure and TX_ERR_DOMAIN_EXCEPTION, which a consumer reads
as a retryable node fault -- deterministic for the same input, so the node never
makes progress. For the utxo-heights and custom-flags arrays, and for
cgo_to_asm, the negative value was cast straight to size_t and produced a span
of about 2^64 elements, which is undefined behaviour rather than a clean error.

Every length parameter, length out-parameter and length return value on this
ABI is now uint64_t, and the shims build their spans through guarded helpers
that reject a pointer/length pair they cannot honour. No size ceiling is
introduced anywhere: a consensus-valid transaction with a large aggregate of
previous locking scripts must still validate, so the only rejections added are
for lengths that cannot describe a real buffer.

Such a rejection is reported in a new TxError domain, TX_ERR_DOMAIN_ABI, rather
than as an exception. An ABI rejection is a statement about the call, not a
verdict on the transaction, and a consumer that cannot tell the two apart would
either retry forever or mark a valid transaction invalid. TxError keeps its
layout, so this is additive.

The handle-producing entry point of each independently reachable surface is
renamed -- TxValidator_CreateV2, ValidateBatch_CreateV2, cgo_from_asm_v2 and
cgo_to_asm_v2. C has no name mangling, so without this an archive built before
the change would link cleanly against the new headers and keep reading a 32-bit
length. Nothing on a surface is reachable without its handle, so a stale archive
now fails to link instead of silently misbehaving.

Also fixed in the functions this change rewrites: the malloc/new[] allocator
mismatch in the asm entry points, whose blobs are freed by the caller with
free(); an uninitialised *scriptLen on the cgo_from_asm error path; unchecked
malloc results; a missing overflow check before the batch result allocation;
and the absence of any try/catch in VerifyScript, VerifyScriptWithCustomFlags
and ValidateBatch, where an escaping exception crossed extern "C". The handlers
themselves are non-throwing, so a failure while reporting a failure cannot
escape either.

Validation semantics are unchanged. Any input that returns OK, a DoS verdict or
a script verdict before this change returns exactly the same after it, and the
consensus flag keeps its meaning.

Version goes to 1.3.0: the C ABI change is breaking.

+++ module/gobdk/cgobench/noop.go
@@ -28,6 +28,6 @@ func CallCGoSumBytes(data []byte) C.longlong {
 	// which adds the data copying overhead.
 	return C.C_SumBytes(
 		(*C.uchar)(unsafe.Pointer(&data[0])),
-		C.int(len(data)),
+		C.uint64_t(len(data)),
 	)
 }
