+++ module/gobdk/cgobench/noop.h
@@ -1,7 +1,9 @@
 #ifndef NOOP_H
 #define NOOP_H
 
+#include <stdint.h>
+
 void C_NoOp(void);
-long long C_SumBytes(const unsigned char *data, int len);
+long long C_SumBytes(const unsigned char *data, uint64_t len);
 
 #endif // NOOP_H
